used BigDecimal

+++ backend/src/main/java/internship/paymentSystem/backend/models/Transaction.java
@@ -10,6 +10,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -19,12 +20,12 @@ import java.time.LocalDateTime;
 @ToString(callSuper = true)
 public class Transaction extends TransactionEntity {
 
-    public Transaction(TypeTransactionEnum type, ActionTransactionEnum action, Double amount, Long accountID,
+    public Transaction(TypeTransactionEnum type, ActionTransactionEnum action, BigDecimal amount, Long accountID,
                        StatusEnum status, StatusEnum nextStatus) {
         super(type, action, amount, accountID, status, nextStatus);
     }
 
-    public Transaction(TypeTransactionEnum type, ActionTransactionEnum action, Double amount, Long accountID,
+    public Transaction(TypeTransactionEnum type, ActionTransactionEnum action, BigDecimal amount, Long accountID,
                        Long targetAccountID, String targetIban, StatusEnum status, StatusEnum nextStatus) {
         super(type, action, amount, accountID, targetAccountID, targetIban, status, nextStatus);
     }
